requirements.txt
django-cms
django-taggit>=0.11.2
django-taggit-autosuggest
django-classy-tags>=0.3.4.1
